pyproject.toml
[tool.poetry]
name = "sentinel"
version = "0.2.0"
description = "Autonomous AI pentesting platform with real-time red/blue team agents"
authors = ["Iko"]
readme = "README.md"
packages = [{include = "sentinel", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# === LLM & AI ===
anthropic = "^0.42.0"
openai = "^1.59.0"

# === Knowledge Graph ===
neo4j = "^5.27.0"
graphiti-core = "^0.5.0"

# === Workflow Orchestration ===
temporalio = "^1.9.0"

# === Web & API ===
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
websockets = "^14.0"
httpx = "^0.28.0"
aiohttp = "^3.11.0"

# === Browser Automation ===
playwright = "^1.49.0"

# === Security Tools ===
python-nmap = "^0.7.1"
dnspython = "^2.7.0"

# === Data & Validation ===
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
pgvector = "^0.3.6"

# === Utilities ===
structlog = "^24.4.0"
rich = "^13.9.0"
typer = "^0.15.0"
python-dotenv = "^1.0.1"
tenacity = "^9.0.0"
orjson = "^3.10.0"
jinja2 = "^3.1.4"

# === Existing dependencies ===
beautifulsoup4 = "^4.12.0"
lxml = "^5.3.0"
docker = "^7.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
pre-commit = "^4.0.0"

[tool.poetry.scripts]
sentinel = "sentinel.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
